requirements.txt
accelerate
wheel
bitsandbytes
gradio
huggingface-hub
numpy
Pillow
transformers
sentencepiece
protobuf